+++ src/main/java/org/jetbrains/bsp/bazel/server/impl/JavaBuildServerImpl.java
@@ -5,7 +5,7 @@ import ch.epfl.scala.bsp4j.JavacOptionsParams;
 import ch.epfl.scala.bsp4j.JavacOptionsResult;
 import java.util.concurrent.CompletableFuture;
 import org.jetbrains.bsp.bazel.server.bsp.BazelBspServerRequestHelpers;
-import org.jetbrains.bsp.bazel.server.service.JavaBuildServerService;
+import org.jetbrains.bsp.bazel.server.services.JavaBuildServerService;
 
 public class JavaBuildServerImpl implements JavaBuildServer {
