+++ src/main/java/org/jetbrains/bsp/bazel/server/services/BUILD
@@ -1,7 +1,7 @@
 load("@rules_java//java:defs.bzl", "java_library")
 
 java_library(
-    name = "service",
+    name = "services",
     srcs = glob(["*.java"]),
     visibility = ["//visibility:public"],
     deps = [
