+++ src/main/java/org/jetbrains/bsp/bazel/server/services/JavaBuildServerService.java
@@ -1,18 +1,13 @@
-package org.jetbrains.bsp.bazel.server.service;
+package org.jetbrains.bsp.bazel.server.services;
 
 import ch.epfl.scala.bsp4j.BuildTargetIdentifier;
-import ch.epfl.scala.bsp4j.ScalaMainClassesParams;
-import ch.epfl.scala.bsp4j.ScalaMainClassesResult;
-import ch.epfl.scala.bsp4j.ScalaTestClassesParams;
-import ch.epfl.scala.bsp4j.ScalaTestClassesResult;
-import ch.epfl.scala.bsp4j.ScalacOptionsItem;
-import ch.epfl.scala.bsp4j.ScalacOptionsParams;
-import ch.epfl.scala.bsp4j.ScalacOptionsResult;
+import ch.epfl.scala.bsp4j.JavacOptionsItem;
+import ch.epfl.scala.bsp4j.JavacOptionsParams;
+import ch.epfl.scala.bsp4j.JavacOptionsResult;
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
@@ -24,13 +19,13 @@ import org.jetbrains.bsp.bazel.server.resolvers.ActionGraphResolver;
 import org.jetbrains.bsp.bazel.server.resolvers.TargetsResolver;
 import org.jetbrains.bsp.bazel.server.util.ActionGraphParser;
 
-public class ScalaBuildServerService {
+public class JavaBuildServerService {
 
   private final BazelData bazelData;
   private final TargetsResolver targetsResolver;
   private final ActionGraphResolver actionGraphResolver;
 
-  public ScalaBuildServerService(
+  public JavaBuildServerService(
       BazelData bazelData,
       TargetsResolver targetsResolver,
       ActionGraphResolver actionGraphResolver) {
@@ -39,25 +34,26 @@ public class ScalaBuildServerService {
     this.actionGraphResolver = actionGraphResolver;
   }
 
-  public Either<ResponseError, ScalacOptionsResult> buildTargetScalacOptions(
-      ScalacOptionsParams scalacOptionsParams) {
+  public Either<ResponseError, JavacOptionsResult> buildTargetJavacOptions(
+      JavacOptionsParams javacOptionsParams) {
     List<String> targets =
-        scalacOptionsParams.getTargets().stream()
+        javacOptionsParams.getTargets().stream()
             .map(BuildTargetIdentifier::getUri)
             .collect(Collectors.toList());
 
     Map<String, List<String>> targetsOptions =
-        targetsResolver.getTargetsOptions(targets, "scalacopts");
+        targetsResolver.getTargetsOptions(targets, "javacopts");
+    // TODO(andrefmrocha): Remove this when kotlin is natively supported
     ActionGraphParser actionGraphParser =
         actionGraphResolver.getActionGraphParser(
-            targets, ImmutableList.of(Constants.SCALAC, Constants.JAVAC));
+            targets, ImmutableList.of(Constants.JAVAC, Constants.KOTLINC));
 
-    ScalacOptionsResult result =
-        new ScalacOptionsResult(
+    JavacOptionsResult result =
+        new JavacOptionsResult(
             targets.stream()
                 .flatMap(
                     target ->
-                        collectScalacOptionsResult(
+                        collectJavacOptionsResult(
                             actionGraphParser,
                             targetsOptions.getOrDefault(target, new ArrayList<>()),
                             actionGraphParser.getInputsAsUri(target, bazelData.getExecRoot()),
@@ -66,30 +62,15 @@ public class ScalaBuildServerService {
     return Either.forRight(result);
   }
 
-  public CompletableFuture<ScalaTestClassesResult> buildTargetScalaTestClasses(
-      ScalaTestClassesParams scalaTestClassesParams) {
-    System.out.printf("DWH: Got buildTargetScalaTestClasses: %s%n", scalaTestClassesParams);
-    // TODO(illicitonion): Populate
-    return CompletableFuture.completedFuture(new ScalaTestClassesResult(new ArrayList<>()));
-  }
-
-  public CompletableFuture<ScalaMainClassesResult> buildTargetScalaMainClasses(
-      ScalaMainClassesParams scalaMainClassesParams) {
-    System.out.printf("DWH: Got buildTargetScalaMainClasses: %s%n", scalaMainClassesParams);
-    // TODO(illicitonion): Populate
-    return CompletableFuture.completedFuture(new ScalaMainClassesResult(new ArrayList<>()));
-  }
-
-  private Stream<ScalacOptionsItem> collectScalacOptionsResult(
+  private Stream<JavacOptionsItem> collectJavacOptionsResult(
       ActionGraphParser actionGraphParser,
       List<String> options,
       List<String> inputs,
       String target) {
-    List<String> suffixes = ImmutableList.of(".jar", ".js");
-    return actionGraphParser.getOutputs(target, suffixes).stream()
+    return actionGraphParser.getOutputs(target, ImmutableList.of(".jar", ".js")).stream()
         .map(
             output ->
-                new ScalacOptionsItem(
+                new JavacOptionsItem(
                     new BuildTargetIdentifier(target),
                     options,
                     inputs,
